Refactor Step 4: Introduce versioning for all tables
This is the first top-level versioning step, which includes the
backwards-compatibility step to 'migrate' the unversioned per-module
tables (Dicts) to their versioned counterparts. The change is subtle
because it is only in the codec provided for the table serialization.
Where previously a simple S.Dict codec was sufficient, there is now a
S.record indirection (with the versioning) that is defined wihtin each
module.