Allow table-level versioning w/o changes to Data module
Because the data module has knowledge about the structure of the v0
versions of all tables, it needs to be able to migrate forward every
table from v0 to latest.

Introduce an interface in the modules to provide that forward-migration
function for every table and use it in the Data module. Currently these
implementations are the id function because at the current stage, v0 =
current for all modules. But this change shall allow to evolve modules
individually without any changes required to the Data module.