Directly use Elm-land flags decoding facility
Instead of 'parsing' the incoming local storage to a string and then
performing the actual parsing with elm-serialize later, expose the
elm-serialize-generated json decoder to elm-land and let it do the work.
In order for a good fallback to work in case there is no local storage,
the pure Data decoder needs to be adapted just a little bit for the
empty string case.