Create category patterns from the Book page
If the description filter is a regex, provide a UI to create a category
pattern directly from there, or even simply apply the category to the
listed items.

This introduces effects that need to be communicated from the Filter
component to the hosting page, making it necessary to parametrize the
filter in hosting page's Msg type.