Use Integer index for RawEntry table
Since identity-based ids by hashing the content of entries was removed
in d83674a, there is no need to use String-typed ids for book entries.
Hashing an auto-increment value to artificially create this id is weird.

Use integers as ids for entries instead. This requires a codec migration
on the Data level, since no versioning on the key side of the RawEntry
dict is available.

This change moves the versioning one level up, such that the entire
RawEntriess model can evolve without other versioning in the Data model.
This is directly used by adding an auto-increment value to the
RawEntries model which is used within that exclusively.