Fix date range slider thumbs going off-track
When one range slider is dragged to the maximum and the other against
it, the weight of the outer range slider becomes 0, leaving no room to
display the thumb. This made it be displayed off-track. Introduce a
minum slider with equal to the thumb size to fix this display bug.