Delay forward to Book after import by 10ms
There is no way for pages to receive updates if only the shared model
has changed. If the page model containes derived state depending on the
shared model and that changes, they can get out of sync. This can happen
when importing data and forwarding to the Book, as the filter is not
updated with the new possible date values. Introducing a tiny lag
between storing the newly imported data and the forward to the Book page
fixes this.