Re-architect monthly data aggregation
While working on graphic representations of the monthly aggregations, it
became apparent that the existing Aggregation type needed some
improvement. In particular the unordered nature of the Dict type and the
stringly-identification of values via labels (with collision risks) made
prepping the data for visiualization much harder thatn necessary.

Any aggregation is now simply a list of per-month aggregates (displayed
as rows in the tabular view), while the month-aggregates contain the
column values as an ordered list. They each also contain the label,
which has an inconsistency risk, but I have not found a better way to
model it thus far.

Another significant change has been to move the state of running sums
into the Aggregator type in order to avoid label-based matching into a
Dict of running sums. This has simplified the aggregation processing a
good bit.

Aggregators are additionally grouped with an AggregationType which helps
meaningful grouping in more complicated bar charts.